backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
psycopg2-binary
python-multipart
geoalchemy2
python-jose
passlib[bcrypt]==1.7.4
shapely
bcrypt==3.2.0
geopandas
shapely
ping3==4.0.4
PyYAML
python-dotenv
httpx
mgrs
minio
pillow